Added all Load Tests

+++ src/apl115/mips_cpu_instructions.cpp
@@ -257,7 +257,7 @@ mips_error LH(uint32_t rs, uint32_t rt, uint16_t immed, mips_cpu_impl *state, mi
 	if((rs + (uint32_t)((int16_t)immed)) % 2 != 0){
 		return mips_ExceptionInvalidAddress;
 	}
-	int32_t result = (int32_t)((buffer[0] << 8 | buffer[1] << 0));
+	int32_t result = (int32_t)((int16_t)(((int8_t)buffer[0] << 8 | (uint8_t)buffer[1] << 0)));
 	return mips_cpu_set_register(state, rt, (uint32_t)result);
 }
